Update category Bug fixed

+++ src/com/bookstore/service/CategoryServices.java
@@ -11,6 +11,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.bookstore.dao.BookDAO;
 import com.bookstore.dao.CategoryDAO;
 import com.bookstore.entity.Category;
 
@@ -98,9 +99,16 @@ public class CategoryServices {
 
 	public void deleteCategory() throws ServletException, IOException {
 		int categoryId = Integer.parseInt(request.getParameter("id"));
-		categoryDAO.delete(categoryId);
-		
-		String message = "The Category with Id " + categoryId + " has been removed successfully";
+		BookDAO bookDAO = new BookDAO();
+		long numberOfBooks = bookDAO.countByCategory(categoryId);
+		String message;
+		if(numberOfBooks > 0) {
+			message = "Could Not Delete the Category (ID: %d) because it currently contains Some Books";
+			message = String.format(message, numberOfBooks);
+		}else {
+			categoryDAO.delete(categoryId);
+			message = "The Category with Id " + categoryId + " has been removed successfully";
+		}
 		listCategory(message);
 	}
